DVX-7420 Move internal out of pkg
Packages in pkg are supposed to be reused out of spaceflight, while those in internal, are supposed to be reused within spaceflight.

Refs [DVX-7420](https://mydevex.atlassian.net/browse/DVX-7420)
